% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adding_shorthand_PTM_names_to_sequences.R
\name{add_Nterm_mods}
\alias{add_Nterm_mods}
\title{Adds Nterm modifications if any are present at the LHS of a sequence.}
\usage{
add_Nterm_mods(
  df,
  sequence_colname = "sequence_modified",
  modifications_colname = "modifications",
  Nterm_mods_df = short_and_longhand_Nterm_mod_names
)
}
\arguments{
\item{df}{A dataframe object that has at least two columns which describe the peptide sequence and modifications it has respectively.}

\item{sequence_colname}{The column name for strings with peptide sequences.}

\item{modifications_colname}{The column name for strings with descriptions of the PTMs present.}

\item{Nterm_mods_df}{A dataframe with two columns: one for strings which describe the N-terminal modification and the other for the corresponding shorthand string.}
}
\value{
A modified column where all sequences are now 'labelled' if their N-terminus is modified, if appropriate.
}
\description{
Looks at the modifications field of a row and based on the package's inbuilt dataframe, finds its shorthand equivalent. This shorthand text is then appended to the LHS of the sequence in the same row.
}
