% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_positions_of_PTMs_col.R
\name{add_positions_of_PTMs_col}
\alias{add_positions_of_PTMs_col}
\title{Add a new column for residues with PTMs and their positions in their native proteins.}
\usage{
add_positions_of_PTMs_col(df_with_sequences_modified, subtract_1 = FALSE)
}
\arguments{
\item{df_with_sequences_modified}{A dataframe with a column name 'sequence_modified' generated by add_seq_info function.}

\item{subtract_1}{Option to subtract 1 from the start and end columns. Defaults to FALSE.}
}
\value{
The \code{df_with_annotated_seqs} with the new column 'PTMs_and_positions'.
}
\description{
This function requires the add_seq_info function to be run first to add the column called 'sequence_modified'.
From this, and the columns called 'start' and 'end', this function uses this information to fill a new column with
residues with PTMs and their positions in their native proteins. There is the option to subtract 1 from both the start
and end columns because with some proteins the first residue, methionine, is removed after the protein is synthesised.
Therefore, some nomenclatures refer to the 2nd residue as the 1st, which is the case in the protein's mature state.
}
