% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_relative_abundance.R
\name{add_relative_abundance}
\alias{add_relative_abundance}
\title{Create new columns with relative abundances by dividing the abundance for one row by the sum of the total abundace.}
\usage{
add_relative_abundance(df, abundance_colnames)
}
\arguments{
\item{df}{A dataframe with different sample abundance values.}

\item{abundance_colnames}{A vector of column names.}
}
\value{
The original df with new columns with relative abundance values.
}
\description{
If the abundance_colnames parameter is left blank then all column headers with 'abundance' in them will be chosen.
New columns generated will have the same name as the ones they're derived from except they'll have '_relative' appended at the end.
}
