% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adding_shorthand_PTM_names_to_sequences.R
\name{add_seq_info}
\alias{add_seq_info}
\title{Adds a new column with modified sequences that include relevant PTMs called "sequence_modfified".}
\usage{
add_seq_info(
  df_with_sequences_and_modifications,
  residues_to_mark_as_unmodified,
  add_Nterm_mods = F
)
}
\arguments{
\item{df_with_sequences_and_modifications}{An optional parameter. If you don't want to import a dataframe directly from an Excel sheet but have an R dataframe instead, enter the object's name here.}

\item{residues_to_mark_as_unmodified}{An optional parameter. If you enter the shorthand code of a residue, e.g. "K", all sequences will show "Kunmod".}

\item{add_Nterm_mods}{Option to label N-terminally modified sequences if indicated in the modifications field. Uses the short_and_longhand_Nterm_mod_names inbuilt dataframe. Defaults to FALSE.}
}
\value{
Dataframe with a new column called 'sequence_modified'.
}
\description{
Note that for the moment it is not possible to change the title names. The title names must be "modifications" and "sequence". Sheet name does not need quotation marks.
}
