% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adding_shorthand_PTM_names_to_sequences.R
\name{add_unmod_qualifier}
\alias{add_unmod_qualifier}
\title{Labels unmodified residues for PTM labelled sequences in a dataframe.}
\usage{
add_unmod_qualifier(
  df_with_modified_seqs,
  colname_with_modified_seqs = "sequence_modified",
  residues
)
}
\arguments{
\item{df_with_modified_seqs}{A dataframe object that has a column with shortand peptide sequences and PTMs.}

\item{colname_with_modified_seqs}{The column name for strings with shortand peptide sequences and PTMs. Given as a string. Defaults to "sequence_modified".}

\item{residues}{What residues should be marked as 'un'. Can be inputted as a single character, e.g. "K" or as a vector.}
}
\value{
A modified column where all strings are modified with specified residues that don't have any PTMs marked as unmodified.
}
\description{
For a given column name and residues, all strings are modified where specified residues that with no PTMs are marked as unmodified. . E.g. "K" would become "Kun" while "Kac" would stay the same. This function should only be applied after "annotate_seq".
}
