% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_freq_of_PTM_per_seq.R
\name{average_freq_of_PTM_per_seq}
\alias{average_freq_of_PTM_per_seq}
\title{For a given string calculate the average number of that string per row, weighted by abundance or just count.}
\usage{
average_freq_of_PTM_per_seq(
  df_with_annotated_seqs,
  PTMs_of_interest,
  weight_by_abundance = T,
  abundance_colnames,
  decimal_places = 3
)
}
\arguments{
\item{df_with_annotated_seqs}{A dataframe with a column name 'sequence_modified' generated by add_seq_info function}

\item{PTMs_of_interest}{Shorthand form of PTMs, PTMs and residues or just residues.}

\item{weight_by_abundance}{A dataframe of all the longhand and corresponding shorthand names for PTMs.}

\item{abundance_colnames}{A vector of the abundance columns. If this is left blank all column names with 'abundance' in them will be used.}

\item{decimal_places}{How many decimal places the output should be given to.}
}
\value{
A dataframe with the average number of the specified string per row, given by PTMs_of_interest.
}
\description{
It is recommended to use the merge_repeats function before this.
}
