% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_no_of_PTMs.R
\name{Avg_no_of_PTMs}
\alias{Avg_no_of_PTMs}
\title{Function to calculate the average number of PTMs per sequence, weighted by abundance.}
\usage{
Avg_no_of_PTMs(df_with_annotated_seqs, all_PTMs, abundance)
}
\arguments{
\item{df_with_annotated_seqs}{A dataframe with at least two columns: one with modified sequences that have shorthand PTMs and another with abundance.}

\item{all_PTMs}{A vector with all the PTMs of interest. E.g. c("Kcr", "Kac").}

\item{abundance}{The title of the column for abundance, given as a string. If this parameter is left empty}
}
\description{
Function to calculate the average number of PTMs per sequence, weighted by abundance.
}
