% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_rows_with_same_elements_by_colnames.R
\name{delete_rows_with_same_elements_in_cols}
\alias{delete_rows_with_same_elements_in_cols}
\title{Delete rows that have the same values for a vector of specified column names.}
\usage{
delete_rows_with_same_elements_in_cols(
  df,
  vector_of_colnames,
  ignore_rows_where_all_abundances_NA = F
)
}
\arguments{
\item{df}{Any dataframe with column names.}

\item{vector_of_colnames}{Column names the user wishes to group rows by. If there is more than one row per group, these rows will be removed.}

\item{ignore_rows_where_all_abundances_NA}{Option to ignore any rows that have all 'NA' values in column names with 'abundance' in them. Defaults to FALSE.}
}
\value{
The inputted \code{df} with rows removed as dictated by \code{vector_of_colnames}.
}
\description{
This function groups rows that have the same values for all the specified columns.
In addition to the primary function, it's also possible to specifiy if some species are not quantified
to exclude these species from being deleted. This works because if two rows have 'NA' values in all columns with 'abundance'
in it they will be retained. By default ignore_rows_where_all_abundances_NA, the parameter that controls this, is set to true.
}
