% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_repeats.R
\name{merge_repeats}
\alias{merge_repeats}
\title{If two rows have the same sequence and modifications fields, this function will merge them.}
\usage{
merge_repeats(df_with_modifications_and_sequence_cols)
}
\arguments{
\item{df_with_modifications_and_sequence_cols}{A dataframe with two columns called 'modifications' and 'sequence'.}
}
\value{
Modified version of \code{df_with_modifications_and_sequence}.
}
\description{
If two rows have the same sequence and modifications fields, this function will merge them.
}
