% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgf_to_datatable.R
\name{mgf_to_datatable}
\alias{mgf_to_datatable}
\title{Translate .mgf data into a datatable format.}
\usage{
mgf_to_datatable()
}
\value{
The .mgf file as a datatable.
}
\description{
This function converts .mgf files to a tabletable that can be manipulated in R.
}
\details{
Upon running this function a file explorer window will open, allowing the user to select the .mgf files of choice.
}
