% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_to.R
\name{normalise_to}
\alias{normalise_to}
\title{Creates new columns that are normalised to a specified column with abundance values.}
\usage{
normalise_to(df, reference_col, comparison_cols)
}
\arguments{
\item{df}{A dataframe with different sample abundance values.}

\item{reference_col}{The column name of the abundance values that will become the 'reference values', given as a string.}

\item{comparison_cols}{A vector of column names to normalise.}
}
\value{
The original df with new columns with normalised abundance values.
}
\description{
Creates new columns that are normalised to a specified column with abundance values.
}
