% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHAME-package.R
\docType{package}
\name{PHAME-package}
\alias{PHAME}
\alias{PHAME-package}
\title{PHAME: PHAME - PTM Histone Analysis Made Easier}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: William Thomas \email{william.thomas31@btopenworld.com}

}
\keyword{internal}
