% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_of_rows_with_PTM.R
\name{prop_of_rows_with_PTM}
\alias{prop_of_rows_with_PTM}
\title{For a given string, usually a PTM, calculate the proportion of rows with that string}
\usage{
prop_of_rows_with_PTM(
  df_with_annotated_seqs,
  PTMs_of_interest,
  weight_by_abundance = T,
  abundance_colnames,
  decimal_places = 3
)
}
\arguments{
\item{df_with_annotated_seqs}{A dataframe with a column name 'sequence_modified' generated by add_seq_info function}

\item{PTMs_of_interest}{A vector of shorthand PTM names. For example "Kac".}

\item{weight_by_abundance}{If true, the proportions will be weighted by abundance.}

\item{abundance_colnames}{A vector of the abundance columns. If this is left blank all column names with 'abundance' in them will be used.}

\item{decimal_places}{How many decimal places the output should be given to.}
}
\value{
A dataframe with the average number of the specified string per row, given by PTMs_of_interest.
}
\description{
It is recommended to use the merge_repeats function before this.
}
