% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{short_and_longhand_Nterm_mod_names}
\alias{short_and_longhand_Nterm_mod_names}
\title{N-terminal modification's longhand and shorthand names.}
\format{
A data frame with 2 rows and 2 variables:
\describe{
\item{Nterm_mods_long}{longhand name of N-terminal modification, string}
\item{Nterm_mods_short}{shorthand name of N-terminal modification, string}
}
}
\usage{
short_and_longhand_Nterm_mod_names
}
\description{
A tibble containing a N-terminal modification's longhand and corresponding shorthand name. This is not an exhaustive list and can be modified by the user to suit their needs. This is most easily done by using the 'editData' function.
}
\keyword{datasets}
